pyproject.toml

[project]
python = "^3.14"
name = "eps-repo-status"

[tool.poetry]
name = "eps-repo-status"
version = "0.0.1-alpha"
description = "EPS VPC Resources"
license = "MIT"
authors = [
    "EPS"
]
readme = "README.md"
repository = "https://github.com/NHSDigital/eps-repo-status"
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
pygithub = "^2.9.0"
requests = "^2.33.1"
boto3 = "^1.42.90"

[tool.poetry.scripts]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest = "^9.0.3"
pytest-cov = "^7.1.0"
black = "^26.3.1"
flake8 = "^7.3.0"

[tool.black]
line-length = 120
